remove unused modin.pandas files
